requirements.txt
requests
beautifulsoup4
python-telegram-bot
matplotlib
lmdb
lxml